Restore manager.New signature, initialization code
